ITR data production checkpoint
We have a good (but not perfect) mechanism now for inferring corp floats based on observable market transactions.

A more robust long-term solution (which will pay dividends in other analyses) will be to switch from the DERA summary data (which we currently use) to the detailed dataset (which is 10x larger).  But that's a matter for another day.  See https://github.com/os-climate/DERA-ingest-pipeline/issues/8