Treat tag table same as sub, num
The special-case logic trying to consolidate tag files made incremental updates of the data unnecessarily complex.  Treat it just like the other two tables we ingest.

Also update Pipfile so we can `pipenv install` and go.

We still need to implement better incremental logic, but now it's just a mere matter of programming, not a Whole Thing.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>